+++ priv/repo/migrations/20170728142856_change_default_status_of_idea_to_under_review.exs
@@ -1,0 +1,15 @@
+defmodule Icta.Repo.Migrations.ChangeDefaultStatusOfIdeaToUnderReview do
+  use Ecto.Migration
+
+  def up do
+    alter table(:ideas) do
+      modify :status, :string, default: "under_review", null: false
+    end
+  end
+
+  def down do
+    alter table(:ideas) do
+      modify :status, :string, default: "new", null: false
+    end
+  end
+end