+++ web/models/idea.ex
@@ -20,7 +20,7 @@ defmodule Icta.Idea do
     struct
     |> cast(params, [:title, :body, :user_id, :owner_id, :status])
     |> validate_required([:title, :user_id])
-    |> validate_inclusion(:status, ["new", "planned", "in_progress", "done"])
+    |> validate_inclusion(:status, ["under_review", "denied", "new", "planned", "in_progress", "done"])
   end
 
   def all_with_votes(user) do
@@ -34,7 +34,7 @@ defmodule Icta.Idea do
   end
 
   defp vote_count_query(user) do
-    from i in Icta.Idea,
+    query = from i in Icta.Idea,
       left_join: v_up in Icta.Vote, on: i.id == v_up.idea_id and v_up.vote == true,
       left_join: v_down in Icta.Vote, on: i.id == v_down.idea_id and v_down.vote == false,
       left_join: my_vote in Icta.Vote, on: i.id == my_vote.idea_id and my_vote.user_id == ^user.id,
@@ -46,6 +46,12 @@ defmodule Icta.Idea do
         up: count(v_up.id, :distinct), down: count(v_down.id, :distinct), my_vote: my_vote.vote,
         comments_count: count(comments.id, :distinct), status: i.status},
       group_by: [i.id, i.title, i.body, i.status, user.name, user.id, owner.id, my_vote.vote]
+
+    query = if user.kind == "user" do
+      from i in query, where: ( i.status != "under_review" or i.status != "denied" ) or i.user_id == ^user.id
+    end
+
+    query
   end
 
 end